% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\docType{data}
\name{ALL.MODELS}
\alias{ALL.MODELS}
\title{A character vector of names of all available models}
\format{An object of class \code{character} of length 10.}
\usage{
ALL.MODELS
}
\description{
A character vector of names of all available models
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\keyword{datasets}

