% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{default}
\alias{default}
\title{Create an R environment that contains default values for root parameters}
\usage{
default(student.var = NULL, avg.success = NULL, time = NULL,
  S.st.var = NULL, L.st.var = NULL, bkt.slip.st.var = NULL,
  bkt.guess.st.var = NULL, min.ntree = NULL, min.depth = NULL,
  min.it.per.tree = NULL, per.item = NULL, bkt.mod = NULL,
  density = NULL, alpha.c = NULL, alpha.p = NULL, p.min = NULL,
  abi.mean = NULL, abi.sd = NULL, trans = NULL, ...)
}
\arguments{
\item{student.var}{variance of student expected success rate}

\item{avg.success}{mean value of the response matrix}

\item{time}{the number of time steps for sequential data}

\item{S.st.var}{variance of student expected success rates of the skill matrix}

\item{L.st.var}{variance of student expected success rates of Learning Transition matrix}

\item{bkt.slip.st.var}{variance of expected values of students in Slip vector of BKT model}

\item{bkt.guess.st.var}{variance of expected values of students in Guess vector of BKT model}

\item{min.ntree}{minimum number of connected components of the partial order structure of items}

\item{min.depth}{minimum depth of the connected components of the partial order structure of items}

\item{min.it.per.tree}{minimum number of items per each connected component of the partial order structure}

\item{per.item}{a boolean value indicates if the students can improve after taking each item}

\item{bkt.mod}{a character string indicates which model governs the generating process for sequential data}

\item{density}{a real value between 0 and 1, indicates the connection density of the partial order structure of items}

\item{alpha.c}{parameter for learning by POKS model, see reference}

\item{alpha.p}{parameter for learning by POKS model, see reference}

\item{p.min}{p-value for interaction test while constructing POK structure}

\item{abi.mean}{mean value of the student abilities vector}

\item{abi.sd}{standard deviation of the student abilities vector}

\item{trans}{a boolean value indicates if transitive links are allowed in the partial order structure of items}
}
\value{
an environment containing all default values of the specified parameters
}
\description{
Create an R environment that contains default values for root parameters
}
\examples{
# Example: 
# Declare a context where there are 15 students and 20 items
p <- pars(students = 15, items = 20)
# Add information that the Partial Order Structure should have depth of 3, two connected components and no transitive links
p <- pars(p, min.depth = 3, max.depth = 3, min.ntree = 2, max.ntree = 2, trans = FALSE)
# Generate data to calculate the `po` parameter
poks.data <- gen("poks", p)
# Visualise the Partial Order Structure
v <- viz(poks.data$po)
# Print the analysed structure
print(v)
}
\author{
Hoang-Trieu trinh, \email{thtrieu@apcs.vn}
}

