% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{down.stream}
\alias{down.stream}
\title{Propagate information downwards}
\usage{
down.stream(pars)
}
\arguments{
\item{pars}{an object of \code{context} class describes all available information in current context}
}
\value{
a new \code{context} object obtained from the input
}
\description{
This function takes the available parameters and try to learn all
possible parameters at lower level, simultaneously check for conflicts
}
\details{
This function use breadth-first scheme to propagate information in
the STRUCTURE, using the learning functions indicated in 3rd element
of each node inside STRUCTURE to learn the corresponding values of
nodes indicated in the 1st element
}
\author{
Hoang-Trieu Trinh
}

