% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{gen.apply}
\alias{gen.apply}
\title{Generate data for a set of models}
\usage{
gen.apply(models, pars, multiply = TRUE, n = 1, progress = FALSE)
}
\arguments{
\item{pars}{another param}

\item{n}{another param}

\item{progress}{another one}

\item{model}{a param}
}
\value{
whatever it is
}
\description{
This function does something
}
\details{
more detail huh?
}
\author{
Hoang-Trieu Trinh
}
\seealso{
what?
}

