% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{get.par}
\alias{get.par}
\title{Get a parameter from the current context}
\usage{
get.par(target, pars, progress = FALSE)
}
\arguments{
\item{target}{a character string indicates the target's name}

\item{pars}{an object of \code{context} class describes the given context}

\item{progress}{a boolean value indicates if the generating process should be printed or not}
}
\value{
if success, a list with two components
\item{value}{value of the target}
\item{context}{the corresponding context}
if not success, NULL
}
\description{
This function generates (if needed) the required target from a context
}
\details{
This function uses \code{up.stream} to obtain target's value and context
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\seealso{
\code{up.stream}, \code{gen}
}

