% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\docType{data}
\name{KEEP}
\alias{KEEP}
\title{List of parameters to be kept for each model in the synthesize process}
\format{An object of class \code{list} of length 10.}
\usage{
KEEP
}
\description{
List of parameters to be kept for each model in the synthesize process
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\keyword{datasets}

