% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{learn}
\alias{learn}
\title{Learn the most probable context for a given data}
\usage{
learn(model, data)
}
\arguments{
\item{model}{a character string indicates}

\item{pars}{another param}

\item{n}{another param}

\item{progress}{another one}
}
\value{
whatever it is
}
\description{
Learn the most probable context for a given data
}
\details{
more detail huh?
}
\author{
Hoang-Trieu Trinh
}
\seealso{
what?
}

