% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{pars}
\alias{pars}
\title{Create or update a context}
\usage{
pars(old.pars = NULL, init.vals = list(student.var = 1/12, avg.success =
  0.5, time = 50, S.st.var = 1/12, L.st.var = 1/12, bkt.slip.st.var = 1/12,
  bkt.guess.st.var = 1/12, min.ntree = 1, min.depth = 1, min.it.per.tree = 1,
  per.item = FALSE, bkt.mod = "dina", density = 0.5, alpha.c = 0.25, alpha.p =
  0.25, p.min = 0.5, abi.mean = 0, abi.sd = 1), dis = NULL, dif = NULL,
  abi = NULL, abi.mean = NULL, abi.sd = NULL, st.exp = NULL,
  it.exp = NULL, items = NULL, concepts = NULL, students = NULL,
  state = NULL, po = NULL, or.t = NULL, or.f = NULL,
  student.var = NULL, avg.success = NULL, min.ntree = NULL,
  max.ntree = NULL, min.depth = NULL, max.depth = NULL, density = NULL,
  min.it.per.tree = NULL, max.it.per.tree = NULL, alpha.c = NULL,
  alpha.p = NULL, p.min = NULL, slip = NULL, guess = NULL,
  per.item = NULL, order = NULL, Q = NULL, S = NULL, M = NULL,
  L = NULL, bkt.mod = NULL, S.st.var = NULL, S.con.exp = NULL,
  L.st.var = NULL, L.con.exp = NULL, skill.space.size = NULL,
  skill.space = NULL, skill.dist = NULL, concept.exp = NULL,
  bkt.slip = NULL, bkt.guess = NULL, time = NULL,
  bkt.slip.it.exp = NULL, bkt.slip.st.var = NULL, bkt.guess.it.exp = NULL,
  bkt.guess.st.var = NULL, irt = NULL, exp = NULL, dina = NULL,
  dino = NULL, nmf.con = NULL, nmf.dis = NULL, nmf.com = NULL,
  lin.pes = NULL, lin.avg = NULL, poks = NULL, bkt = NULL)
}
\arguments{
\item{old.pars}{an object of \code{context} class describe the context that needed to be updated,
leave this parameter if the user need a new context.}
}
\value{
an object of \code{context} class describe the updated or new context
}
\description{
This function allows user input a set of parameters to a new or available context
}
\details{
This function takes in a set of parameters that the user input and assemble them
into a \code{context} object, it uses the \code{down.stream} function to check for conflicts
}
\author{
Hoang-Trieu Trinh
}
\seealso{
\code{down.stream}
}

