% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{up.stream}
\alias{up.stream}
\title{Propagate information upwards}
\usage{
up.stream(target, pars, progress = FALSE)
}
\arguments{
\item{target}{a character string indicates the target's name}

\item{pars}{an object of \code{context} class describes all available information in current context}

\item{progress}{a boolean value indicates if the generating process should be printed}
}
\value{
a new \code{context} object obtained from the input, if the target cannot be reach,
the old \code{context} object is returned
}
\description{
This function takes the available parameters and generate higher
level parameters in a tailored direction so that a specified
target can be reach
}
\details{
This function runs a recursive search for available information at lower level
nodes in the STRUCTURE that has been provided by the input. Whenever there is
more than two ways to generate a parameter, the function chooses the one that
requires inputs that is more likely to be learned directly from the target.
}
\author{
Hoang-Trieu Trinh
}
\seealso{
\code{which.closest}
}

