% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\docType{data}
\name{ALL.MODELS}
\alias{ALL.MODELS}
\title{A character vector of names of all available models}
\format{\preformatted{ chr [1:10] "exp" "irt" "poks" "dina" "dino" "lin.avg" "nmf.con" ...
}}
\usage{
ALL.MODELS
}
\description{
A character vector of names of all available models
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\keyword{datasets}

