% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{gen.apply}
\alias{gen.apply}
\title{Generate data for a set of models and contexts}
\usage{
gen.apply(models, pars, multiply = TRUE, n = 1, progress = FALSE)
}
\arguments{
\item{pars}{a context or a list of contexts}

\item{multiply}{a boolean value indicates in which way should \code{models} and \code{pars} be matched.
if \code{TRUE} the generating process will be performed on every possible combination from \code{models} and \code{pars},
if \code{FALSE} each model will be matched with its respective context in the same order specified in \code{models} and \code{pars},
in other words, set \code{multiply} to \code{FALSE} will make \code{gen.apply} does the exact same thing to \code{mapply(gen,models,pars)}}

\item{n}{number of runs for each generation}

\item{progress}{a boolean value indicates if the generating steps should be printed or not.}

\item{model}{a character string indicates which model governs the learning process.}
}
\value{
a matrix with each entry is a context or a list of contexts, depending on the format indicated by \code{multiply}
}
\description{
Generate data for a set of models and contexts
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\seealso{
\code{gen}, \code{mapply}, \code{sapply}
}

