% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{gen}
\alias{gen}
\title{Generate data for a model}
\usage{
gen(model, pars, n = 1, progress = FALSE)
}
\arguments{
\item{model}{a param}

\item{pars}{another param}

\item{n}{another param}

\item{progress}{another one}
}
\value{
whatever it is
}
\description{
This function does something
}
\details{
more detail huh?
}
\author{
Hoang-Trieu Trinh
}
\seealso{
what?
}

