% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{gen}
\alias{gen}
\title{Generate data for a model}
\usage{
gen(model, pars, n = 1, progress = FALSE)
}
\arguments{
\item{model}{a character string indicates which model governs the generating process.}

\item{pars}{a context}

\item{n}{numer of runs}

\item{progress}{a boolean value indicates if the generating steps should be printed or not.}
}
\value{
a context with its data unit activated
}
\description{
This function generates a context with a specified data unit activated
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\seealso{
\code{get.par}
}

