% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get.par}
\alias{get.par}
\title{Get a parameter from the current context}
\usage{
get.par(target, pars, progress = FALSE)
}
\arguments{
\item{target}{a character string indicates the target's name}

\item{pars}{an object of \code{context} class describes the given context}

\item{progress}{a boolean value indicates if the generating process should be printed or not}
}
\value{
if success, a list with two components
\item{value}{value of the target}
\item{context}{the corresponding context}
if not success, NULL
}
\description{
This function generates (if needed) the required target from a context
}
\examples{
# Declare a context
p <- pars(students = 20, items = 15)
# Regular way to access information inside a context
p$students # returns 20
# get.par is a alternative
get.par("students", p)
# However, it is not trivial to generate a skill mastery matrix from p
p$M # NULL returned
# get.par can do this
M <- get.par("M", p, progress = True)
print(M)
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\seealso{
\code{gen}
}

