% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\docType{data}
\name{KEEP}
\alias{KEEP}
\title{List of parameters to be kept for each model in the synthesize process}
\format{\preformatted{List of 10
 $ exp    : chr [1:3] "avg.success" "it.exp" "student.var"
 $ irt    : chr [1:4] "dis" "dif" "abi.mean" "abi.sd"
 $ poks   : chr [1:6] "po" "student.var" "state" "alpha.c" ...
 $ dina   : chr [1:5] "Q" "skill.space" "skill.dist" "slip" ...
 $ dino   : chr [1:5] "Q" "skill.space" "skill.dist" "slip" ...
 $ lin.avg: chr [1:3] "Q" "S.st.var" "S.con.exp"
 $ nmf.con: chr [1:2] "Q" "concept.exp"
 $ nmf.dis: chr [1:2] "Q" "concept.exp"
 $ nmf.com: chr [1:2] "Q" "concept.exp"
 $ bkt    : chr [1:12] "S.st.var" "S.con.exp" "L.st.var" "L.con.exp" ...
}}
\usage{
KEEP
}
\description{
List of parameters to be kept for each model in the synthesize process
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\keyword{datasets}

