% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{learn}
\alias{learn}
\title{Learn the most probable context for a given data}
\usage{
learn(model, data)
}
\arguments{
\item{model}{a character string indicates which model governs the learning process.}

\item{data}{a list contains data that needs to be synthesized, first component is the response matrix, the other components are additional information that the specified model requires.}
}
\value{
the most probable context corresponds to \code{data} under the assumptions made by \code{model}
}
\description{
Learn the most probable context for a given data
}
\examples{
# Let us make up some data to learn
p <- pars(students = 20, concepts = 4, items = 15)
poks.data <- gen("poks", p)
# Notice that the generated data is in component poks,
# together with other hyper-parameters for learning by POKS model.
poks.data$poks
# Assume we want to learn with DINA (Deterministic Input Noisy And) model
# then poks.data$poks is not relevant anymore, instead we need a Q-matrix
# For demonstration, let's just make it up
Q <- get.par("Q", p)$value
R <- poks.data$poks$R
dina_dat <- list(R=R, Q=Q)
# Now learn from poks.data
learned.poks <- learn("dina", data = dina_dat)
# learned.poks is a context with learned parameters
class(learned.poks) # returns "context"
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}

