% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{learn}
\alias{learn}
\title{Learn the most probable context for a given data}
\usage{
learn(model, data)
}
\arguments{
\item{model}{a character string indicates which model governs the learning process.}

\item{data}{a list contains data that needs to be synthesized, first component is the response matrix, the other components are additional information that the specified model requires.}
}
\value{
the most probable context corresponds to \code{data} in light of \code{model}
}
\description{
Learn the most probable context for a given data
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}

