% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{syn}
\alias{syn}
\title{Generate synthetic data}
\usage{
syn(model, data, keep.pars = keep(model), students = ncol(data$R), n = 1,
  progress = FALSE)
}
\arguments{
\item{model}{a character string indicates which model governs the synthesizing process.}

\item{data}{a list contains data that needs to be synthesized, first component 
is the response matrix, the other components are additional information that the
specified model requires.}

\item{keep.pars}{a character string vector contains names of the parameters to be
 kept after learning the most probable context from \code{data},
by default this parameter is set by values indicated in vector \code{KEEP}.}

\item{students}{number of students in the synthetic data, by default this number is kept.}

\item{n}{number of synthetic dataset(s) to generate.}

\item{progress}{a boolean value indicates if the generating steps should be printed or not.}
}
\value{
a list with two components, first component is identical to argument \code{data}, second is a context or a list of contexts that have been generated.
}
\description{
This function synthesize a given data under the assumptions of a given model
}
\details{
This function is essentially a wrapper of function \code{learn} and \code{gen}, 
where in between it eliminates all parameters that are not indicated in \code{keep.pars} in the learned context.
you are also given the option of changing the number of students.
}
\examples{
# Let us make up some data to synthesize
dina.data <- gen('dina', pars(students = 20, concepts = 4, items = 5))$dina
dina.syn <- syn('dina', data = dina.data, students = 50, n = 10)
}
\author{
Hoang-Trieu Trinh, \email{thtrieu@apcs.vn}
}
\seealso{
\code{learn}, \code{gen}, \code{KEEP}
}

