% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/structure.R
\name{syn}
\alias{syn}
\title{Generate synthetic data from a given data
This function does something}
\usage{
syn(model, data, keep.pars = keep(model), students = ncol(data$R), n = 1,
  progress = FALSE)
}
\arguments{
\item{model}{a param}

\item{n}{another param}

\item{progress}{another one}

\item{pars}{another param}
}
\value{
whatever it is
}
\description{
Generate synthetic data from a given data
This function does something
}
\details{
more detail huh?
}
\author{
Hoang-Trieu Trinh
}
\seealso{
what?
}

